moved 'eda' classes to 'driven _dynamics'
